style files are moved to styles folder, PopUserExitButton style is fixed
