+++ skypro-kanban/src/styles/Global.styled.js
